Ignition 1 refactored
Initial work on refactoring ignition schedules

+++ speeduino/globals.ino
@@ -107,7 +107,7 @@ bool channel6InjEnabled = false;
 bool channel7InjEnabled = false;
 bool channel8InjEnabled = false;
 
-int16_t ignition1EndAngle = 0;
+int ignition1EndAngle = 0;
 int ignition2EndAngle = 0;
 int ignition3EndAngle = 0;
 int ignition4EndAngle = 0;
